% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{clip_wavelogger}
\alias{clip_wavelogger}
\title{clip wavelogger table by date}
\usage{
clip_wavelogger(x, startstop = NA)
}
\arguments{
\item{x}{tibble, waveloger}

\item{startstop}{POSIXt vector of two values or NA, only used if clip = "user"}
}
\value{
tibble
}
\description{
clip wavelogger table by date
}
