% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{example_filepath}
\alias{example_filepath}
\title{retrieve example type wavelogger directory path}
\usage{
example_filepath()
}
\value{
character
}
\description{
retrieve example type wavelogger directory path
}
