% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{interp_swpressure}
\alias{interp_swpressure}
\title{interpolate air pressure to match owhl, calculate seawater pressure}
\usage{
interp_swpressure(
  wavelogger = read_wavelogger(),
  airpressure = read_airpressure()
)
}
\arguments{
\item{wavelogger}{tibble, wavelogger data}

\item{airpressure}{tibble, airpressure data}
}
\value{
tibble
}
\description{
interpolate air pressure to match owhl, calculate seawater pressure
}
