% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{interp_wave_air}
\alias{interp_wave_air}
\title{interpolate sea level pressure to match owhl data}
\usage{
interp_wave_air(
  wavelogger = read_wavelogger(),
  airpressure = read_airpressure()
)
}
\arguments{
\item{wavelogger}{tibble, wavelogger data}

\item{airpressure}{tibble, airpressure data}
}
\value{
tibble
}
\description{
interpolate sea level pressure to match owhl data
}
