% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mbar_to_elevation}
\alias{mbar_to_elevation}
\title{convert pressure to sea surface elevation, correct for signal attenuation}
\usage{
mbar_to_elevation(wavelogger = interp_swpressure(), latitude = 44.5)
}
\arguments{
\item{wavelogger}{tibble, wavelogger data}

\item{latitude}{numeric, approx latitude of deployment - degrees north, default 44.5}
}
\value{
tibble
}
\description{
convert pressure to sea surface elevation, correct for signal attenuation
}
