% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_airpressure_old}
\alias{read_airpressure_old}
\title{read sea level pressure data file}
\usage{
read_airpressure_old(filename = example_airpressure())
}
\arguments{
\item{filename}{character, the name of the air pressure file}
}
\value{
tibble
}
\description{
read sea level pressure data file
}
