% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_wavelogger}
\alias{read_wavelogger}
\title{read wavelogger data file}
\usage{
read_wavelogger(filepath = example_filepath())
}
\arguments{
\item{filepath}{character, the name of the directory - full path needed}
}
\value{
tibble
}
\description{
read wavelogger data file
}
