% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{wavespec_plot}
\alias{wavespec_plot}
\title{Plot waveheight data}
\usage{
wavespec_plot(
  x = wave_stats(),
  main = "Significant Wave Height",
  xlabel = "Date",
  ylabel = "Significant Wave Height (m)",
  facet = NULL,
  ...
)
}
\arguments{
\item{x}{tibble of wave stats data}

\item{main}{character, title}

\item{xlabel}{character, title of xaxis}

\item{ylabel}{character, title of yaxis}

\item{facet}{character, name of the column to facet upon (like "Site") or NULL to skip}

\item{...}{further arguments passed to \code{\link[ggplot2]{theme}}}
}
\value{
ggplot2 object
}
\description{
Plot waveheight data
}
