% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_wavelogger}
\alias{read_wavelogger}
\title{read wavelogger data file}
\usage{
read_wavelogger(filepath = example_filepath(), deploy = NA, recover = NA)
}
\arguments{
\item{filepath}{character, the name of the directory - full path needed}

\item{deploy}{POSIXt or NA, if not NA, clip data before this time}

\item{recover}{POSIXt or NA, if not NA, clip data before this time}
}
\value{
tibble
}
\description{
read wavelogger data file
}
